refactor: Diretório de referencia por conta da sua nomenclatura "Add" para "Post"
