+++ src/modules/users/presentation/components/add-user-menu/add-user-menu-root.component.tsx
@@ -1,18 +1,32 @@
+'use client'
+
 import { DrawerDialog } from '@/modules/shared/presentation/components/dialog-with-drawer'
-import type { ReactNode } from 'react'
-import { useDialog } from './add-user-menu-provider.component'
+import { useEffect, type ReactNode } from 'react'
+import { usePostUserForm } from '../../hooks/use-post-user-form.hook'
+import { FormProvider } from 'react-hook-form'
 
 interface AddUserMenuRootComponentProps {
   children: ReactNode
+  isOpen: boolean
+  close: () => void
 }
 
 export function AddUserMenuRootComponent({
-  children
+  children,
+  isOpen,
+  close
 }: AddUserMenuRootComponentProps) {
-  const { isOpen, close } = useDialog()
+  const { methods, defaultValues } = usePostUserForm()
+
+  useEffect(() => {
+    if (isOpen) methods.reset(defaultValues)
+  }, [isOpen, defaultValues, methods])
+
   return (
-    <DrawerDialog.Root open={isOpen} onOpenChange={close}>
-      {children}
-    </DrawerDialog.Root>
+    <FormProvider {...methods}>
+      <DrawerDialog.Root open={isOpen} onOpenChange={close}>
+        {children}
+      </DrawerDialog.Root>
+    </FormProvider>
   )
 }
