+++ src/modules/users/presentation/components/add-user-menu/add-user-menu-trigger.component.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from '@/modules/shared/presentation/components/shadcn/button'
 import { UserRoundPlus } from 'lucide-react'
-import { useAddUserMenuTrigger } from '../../hooks/use-add-user-menu-trigger.hook'
+import { useAddUserMenuTrigger } from '../../hooks/use-post-user-menu-trigger.hook'
 
 export function AddUserMenuTriggerComponent() {
   const { loadUserAddOpenDialog } = useAddUserMenuTrigger()
