+++ src/modules/users/presentation/components/add-user-menu/add-user-menu.component.tsx
@@ -1,12 +1,11 @@
 'use client'
 
 import { UserForm } from '../user-form'
-import { AddUserForm } from '../add-user-form-provider'
 import { Button } from '@/modules/shared/presentation/components/shadcn/button'
 import { useDialog } from './add-user-menu-provider.component'
 import { AddUserMenu } from '.'
-import { useAddUserSubmit } from '../../hooks/use-add-user-submit.hook'
-import type { UserWithFiles } from '@/modules/users/domain/types/user-with-files'
+import { useAddUserSubmit } from '../../hooks/use-post-user-submit.hook'
+import type { UserEntity } from '@/modules/users/domain/entities/user.entity'
 
 interface AddUserMenuComponentProps {
   title: string
@@ -21,35 +20,33 @@ export function AddUserMenuComponent({
   const { onAction } = useAddUserSubmit()
 
   return (
-    <AddUserMenu.Root>
+    <AddUserMenu.Root isOpen={isOpen} close={close}>
       <AddUserMenu.Content>
         <AddUserMenu.Header title={title} description={description} />
 
-        <AddUserForm.Provider isOpen={isOpen}>
-          <UserForm.Form>
-            <UserForm.Input.Name require />
-            <UserForm.Input.Email require />
-            <UserForm.Input.Company />
-            <UserForm.Input.Position />
-            <UserForm.Input.Username require />
-            <UserForm.Input.PasswordRegDeadline />
-            <UserForm.Input.Files />
-            <UserForm.Input.Description />
-          </UserForm.Form>
+        <UserForm.Form>
+          <UserForm.Input.Name require />
+          <UserForm.Input.Email require />
+          <UserForm.Input.Company />
+          <UserForm.Input.Position />
+          <UserForm.Input.Username require />
+          <UserForm.Input.PasswordRegDeadline />
+          <UserForm.Input.Files />
+          <UserForm.Input.Description />
+        </UserForm.Form>
 
-          <AddUserMenu.Footer>
-            <Button
-              className="w-full sm:w-[150px]"
-              variant="outline"
-              onClick={close}
-            >
-              Cancelar
-            </Button>
-            <UserForm.Submit
-              onSubmit={(user: UserWithFiles) => onAction(user, close)}
-            />
-          </AddUserMenu.Footer>
-        </AddUserForm.Provider>
+        <AddUserMenu.Footer>
+          <Button
+            className="w-full sm:w-[150px]"
+            variant="outline"
+            onClick={close}
+          >
+            Cancelar
+          </Button>
+          <UserForm.Submit
+            onSubmit={(user: UserEntity) => onAction(user, close)}
+          />
+        </AddUserMenu.Footer>
       </AddUserMenu.Content>
     </AddUserMenu.Root>
   )
